+++ joss-paper/paper.bib
@@ -28,7 +28,7 @@ volume = {189},
 pages = {122357},
 year = {2023},
 issn = {0040-1625},
-doi = {https://doi.org/10.1016/j.techfore.2023.122357},
+doi = {10.1016/j.techfore.2023.122357},
 url = {https://www.sciencedirect.com/science/article/pii/S0040162523000422},
 author = {Lara Agnoli and Erin Urquhart and Nikolaos Georgantzis and Blake Schaeffer and Richard Simmons and Bilqis Hoque and Merrie Beth Neely and Claire Neil and Jacques Oliver and Andrew Tyler},
 keywords = {Water quality, Earth observation, Satellite remote sensing, Knowledge transfer, Technology transfer, Data accessibility}
@@ -78,6 +78,7 @@ abstract = {This paper provides a review and summary status of the research unde
     year = {2024},
     note = {R package version 1.7-78},
     url = {https://CRAN.R-project.org/package=terra},
+    doi = {10.32614/CRAN.package.terra}
   }
 
   @Manual{shiny,
@@ -86,6 +87,7 @@ abstract = {This paper provides a review and summary status of the research unde
     year = {2024},
     note = {R package version 1.8.1.1},
     url = {https://CRAN.R-project.org/package=shiny},
+    doi = {10.32614/CRAN.package.shiny}
   }
 
   @Article{tidyverse,
@@ -126,6 +128,7 @@ abstract = {This paper provides a review and summary status of the research unde
          author = "Koen Hufkens",
          year = "2022",
          url = "https://github.com/ropensci/MODISTools"
+         doi = "10.32614/CRAN.package.MODISTools"
          }
 
 @Article{MODIStsp,
@@ -172,7 +175,7 @@ pages = {18-27},
 year = {2017},
 note = {Big Remotely Sensed Data: tools, applications and experiences},
 issn = {0034-4257},
-doi = {https://doi.org/10.1016/j.rse.2017.06.031},
+doi = {10.1016/j.rse.2017.06.031},
 url = {https://www.sciencedirect.com/science/article/pii/S0034425717302900},
 author = {Noel Gorelick and Matt Hancher and Mike Dixon and Simon Ilyushchenko and David Thau and Rebecca Moore},
 keywords = {Cloud computing, Big data, Analysis, Platform, Data democratization, Earth Engine},
