+++ joss-paper/paper.md
@@ -34,7 +34,7 @@ Data from Earth Observation satellites are a crucial and increasingly valuable r
 
 Despite the increasing availability and utility of these EO data, accessing and using them presents several challenges. Researchers often encounter issues such as multiple data sources, data complexity, or large file sizes [@AGNOLI2023122357]. These data are typically distributed as multidimensional layers over extensive areas, making accessibility and processing difficult, especially when large time series are required. This problem is particularly acute in developing regions where internet infrastructure, needed to access the data, can be limited. The complexity of accessing EO data can result in obstacles such as underutilization or the creation of siloed data processing workflows, which separate data extraction from pre-processing and analysis, thereby hindering transparent and reproducible open science practices. While some powerful tools, such as [Google Earth Engine](https://earthengine.google.com/) [@GORELICK201718] offer some solutions to these problems, they also have important limitations, such as proprietary and energy-intensive software. Altogether, these barriers hinder the full potential of Earth observation data to support global research and decision-making. Efforts to simplify and streamline access to these data, while maintaining an open-source and open-science framework, are essential to overcoming these obstacles and maximizing the benefits of satellite data for all.
 
-Here, we introduce `modisfast`, an open-source R package [@R] designed to simplify, streamline, and accelerate the download and import of MODIS, VIIRS, and GPM time series for R users. This package expands and the existing ecosystem of R tools for accessing MODIS data, enhancing it by introducing new features and data sources (see section \autoref(comp-other-soft)). `modisfast` allows users to subset these datasets using spatial, temporal, and band/layer directly at the downloading phase, optimizing data download and processing. Additionally, downloads can be parallelized for increased efficiency. `modisfast` thus facilitates access to EO data for R users, particularly in regions with limited internet infrastructure, and enables embedding data extraction within complex and holistic data workflows in R - fostering transparency and reproducibility in the context of Open Science. Importantly, the foundational framework of `modisfast` (see section \autoref(foundational-fmwrk)) guarantees the package's long-term sustainability, open-source nature, and cost-free availability.
+Here, we introduce `modisfast`, an open-source R package [@R] designed to simplify, streamline, and accelerate the download and import of MODIS, VIIRS, and GPM time series for R users. This package expands and the existing ecosystem of R tools for accessing MODIS data, enhancing it by introducing new features and data sources (see section \autoref{comp-other-soft}). `modisfast` allows users to subset these datasets using spatial, temporal, and band/layer directly at the downloading phase, optimizing data download and processing. Additionally, downloads can be parallelized for increased efficiency. `modisfast` thus facilitates access to EO data for R users, particularly in regions with limited internet infrastructure, and enables embedding data extraction within complex and holistic data workflows in R - fostering transparency and reproducibility in the context of Open Science. Importantly, the foundational framework of `modisfast` (see section \autoref{foundational-fmwrk}) guarantees the package's long-term sustainability, open-source nature, and cost-free availability.
 
 # Target audience
 
@@ -91,9 +91,9 @@ The full list of available data collections can be accessed with the function `m
 
[…]
