### 2024-07-18 6f23d7d
update bib for joss paper
