+++ joss-paper/paper.bib
@@ -127,7 +127,7 @@ abstract = {This paper provides a review and summary status of the research unde
          title = "The MODISTools package: an interface to the MODIS Land Products Subsets Web Services",
          author = "Koen Hufkens",
          year = "2022",
-         url = "https://github.com/ropensci/MODISTools"
+         url = "https://github.com/ropensci/MODISTools",
          doi = "10.32614/CRAN.package.MODISTools"
          }
 
