### 2024-07-19 #7006 @ptaconet
@kthyng Hello ! I have made the modifications to the pdf, it seems to work now.

Potential reviewers for this paper could be any person from the proposed [list of person](https://reviewers.joss.theoj.org/lookup) matching with the following query : Language: R ; Area: Earth Sciences and Ecology ; Keywords: remote sensing   (https://reviewers.joss.theoj.org/reviewers/search?language=29&area_id=6&keywords=remote+sensing&name=&commit=Update+search)

Additional reviewers might include : [Annelise Tran](https://orcid.org/0000-0001-5463-332X), [Jean-Baptise Ferret](https://jbferet.gitlab.io/), Emmanuel Roux(https://orcid.org/0000-0003-2266-8207), Vincent Herbreteau(https://orcid.org/0000-0003-1762-6336)
