### 2024-07-22 #7006 @kthyng
@ptaconet I am going to ping the greater editorial board to take a look at this package to see if it is in scope in terms of [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) due to its relatively small size. This process will take a few weeks; thanks for your patience.
