### 2024-08-06 d14bdcf
add code of conduct and contributing
