+++ README.md
@@ -30,14 +30,32 @@ provides the MOD11A1.006 (MODIS/Terra Land Surface Temperature/Emissivity Daily 
 This package enables to build OPeNDAP (https) URLs given input parameters such as a data collection, region and time range of interst . These URLs can then be used to either download the data to your workspace or computer, or access the datacube directly as an R object (of class `ndcf4`, `raster`, `stars`, etc.)
 -->
 
-**`modisfast`** (formerly `opendapr`) is an R package that provides
-functions to **speed-up** the **download** of time-series raster data
-products derived from some
+## Table of contents
+
+<p align="left">
+• <a href="#overview">Overview</a><br> •
+<a href="#installation">Installation</a><br> •
+<a href="#get-started">Get started</a><br> •
+<a href="#collections-available-in-modisfast">Data collections
+available</a><br> • <a href="#foundational-framework">Foundational
+framework </a><br> •
+<a href="#comparison-with-similar-r-packages">Comparison with similar R
+packages</a><br> • <a href="#future-developments">Future
+developments</a><br> • <a href="#citation">Citation</a><br> •
+<a href="#contributing">Contributing</a><br> •
+<a href="#acknowledgments">Acknowledgments</a><br>
+</p>
+
+## Overview
+
+**`modisfast`** is an R package designed for **easy** and **fast**
+downloads and import of some widely-used satellite-derived environmental
+data, including
 [**MODIS**](https://lpdaac.usgs.gov/data/get-started-data/collection-overview/missions/modis-overview/)
-and
+Land products,
 [**VIIRS**](https://lpdaac.usgs.gov/data/get-started-data/collection-overview/missions/s-npp-nasa-viirs-overview/)
-observations, as well as other widely-used satellite-derived
-environmental data (e.g. Global Precipitation Measurement Mission).
+Land products, and \[**GPM**\])(<https://gpm.nasa.gov/data/>) products
+(Global Precipitation Measurement Mission).
 
 **`modisfast`** uses the abilities offered by the [OPeNDAP
 framework](https://www.opendap.org/about/) (*Open-source Project for a
@@ -53,15 +71,16 @@ data **over long time series** and **over areas**, rather than short
 time series and points.
 
 Below is a comparison of modisfast with other packages available for
-downloading chunks of remote sensing data :
+downloading chunks of MODIS or VIIRS data :
 
-| Package                                                  |                   Data                    | Temporal subsetting\* | Spatial subsetting\* | Dimensional subsetting\* |       Speed        |
-|:---------------------------------------------------------|:-----------------------------------------:|:---------------------:|:--------------------:|:------------------------:|:------------------:|
-| [`modisfast`](https://github.com/ptaconet/modisfast)     |             MODIS, VIIRS, GPM             |  :white_check_mark:   |  :white_check_mark:  |    :white_check_mark:    | :white_check_mark: |
-| [`MODIS`](https://github.com/fdetsch/MODIS)              |                   MODIS                   |  :white_check_mark:   |         :x:          |           :x:            | :white_check_mark: |
-| [`MODIStsp`](https://github.com/ropensci/MODIStsp)       |                   MODIS                   |  :white_check_mark:   |         :x:          |    :white_check_mark:    | :white_check_mark: |
-| [`MODISTools`](https://github.com/ropensci/MODISTools)   | MODIS and VIIRS (but not all collections) |  :white_check_mark:   |  :white_check_mark:  |    :white_check_mark:    | :white_check_mark: |
-| [`appeears`](https://github.com/bluegreen-labs/appeears) |       MODIS, VIIRS, and many others       |  :white_check_mark:   |  :white_check_mark:  |    :white_check_mark:    |        :x:         |
[…]
