+++ README.Rmd
@@ -40,22 +40,42 @@ https://opendap.cr.usgs.gov/opendap/hyrax/MOD11A1.006/h17v08.ncml.nc4?MODIS_Grid
 provides the MOD11A1.006 (MODIS/Terra Land Surface Temperature/Emissivity Daily L3 Global 1km SIN Grid V006) data in netCDF, subsetted for bands LST_Day_1km, LST_Night_1km, QC_Day, QC_Night, for each day between the 2017-01-01 and the 2017-01-30, and within the following bounding box (lon/lat): -5.41 8.84, -5.82 9.54.
 This package enables to build OPeNDAP (https) URLs given input parameters such as a data collection, region and time range of interst . These URLs can then be used to either download the data to your workspace or computer, or access the datacube directly as an R object (of class `ndcf4`, `raster`, `stars`, etc.)
 -->
-**`modisfast`** (formerly `opendapr`) is an R package that provides functions to **speed-up** the **download** of time-series raster data products derived from some [**MODIS**](https://lpdaac.usgs.gov/data/get-started-data/collection-overview/missions/modis-overview/) and [**VIIRS**](https://lpdaac.usgs.gov/data/get-started-data/collection-overview/missions/s-npp-nasa-viirs-overview/) observations, as well as other widely-used satellite-derived environmental data (e.g. Global Precipitation Measurement Mission).
+
+## Table of contents
+
+<p align="left">
+  • <a href="#overview">Overview</a><br>
+  • <a href="#installation">Installation</a><br>
+  • <a href="#get-started">Get started</a><br>
+  • <a href="#collections-available-in-modisfast">Data collections available</a><br>
+  • <a href="#foundational-framework">Foundational framework </a><br>
+  • <a href="#comparison-with-similar-r-packages">Comparison with similar R packages</a><br>
+  • <a href="#future-developments">Future developments</a><br>
+  • <a href="#citation">Citation</a><br>
+  • <a href="#contributing">Contributing</a><br>
+  • <a href="#acknowledgments">Acknowledgments</a><br>
+</p>
+
+## Overview
+
+**`modisfast`** is an R package designed for **easy** and **fast** downloads and import of some widely-used satellite-derived environmental data, including  [**MODIS**](https://lpdaac.usgs.gov/data/get-started-data/collection-overview/missions/modis-overview/) Land products, [**VIIRS**](https://lpdaac.usgs.gov/data/get-started-data/collection-overview/missions/s-npp-nasa-viirs-overview/) Land products, and [**GPM**])(https://gpm.nasa.gov/data/) products (Global Precipitation Measurement Mission).
 
 **`modisfast`** uses the abilities offered by the [OPeNDAP framework](https://www.opendap.org/about/) (*Open-source Project for a Network Data Access Protocol*) to download a subset of Earth science data cube, along spatial, temporal or any other data dimension (depth, ...). This way, it reduces downloading time and disk usage to their minimum : no more 1° x 1° MODIS tiles with 10 bands when your region of interest is only 30 km x 30 km wide and you need 2 bands ! Moreover, `modisfast` enables parallel downloads of data.
 
 `modisfast` is hence particularly suited for retrieving MODIS or VIIRS data **over long time series** and **over areas**, rather than short time series and points.
 
-Below is a comparison of modisfast with other packages available for downloading chunks of remote sensing data :
+Below is a comparison of modisfast with other packages available for downloading chunks of MODIS or VIIRS data :
 
 
-|    Package                                           | Data  |  Temporal subsetting* | Spatial subsetting* | Dimensional subsetting*   | Speed |
[…]
