### 2024-08-07 d10e61d
add vignette for package comparison with other packages
