+++ joss-paper/paper.bib
@@ -81,14 +81,6 @@ abstract = {This paper provides a review and summary status of the research unde
     doi = {10.32614/CRAN.package.terra}
   }
 
-  @Manual{shiny,
-    title = {shiny: Web Application Framework for R},
-    author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
-    year = {2024},
-    note = {R package version 1.8.1.1},
-    url = {https://CRAN.R-project.org/package=shiny},
-    doi = {10.32614/CRAN.package.shiny}
-  }
 
   @Article{tidyverse,
     title = {Welcome to the {tidyverse}},
