### 2024-08-09 #7006 @kthyng
@ptaconet Please shorten your paper. You can move material to your docs as needed, for example. Aim for 1000 words max, which you can check with `@editorialbot check repository`.
