### 2024-08-26 4fb4781
update JOSS paper and website
