+++ joss-paper/paper.md
@@ -30,33 +30,29 @@ bibliography: paper.bib
 
 # Statement of need
 
-Data from Earth Observation satellites are a crucial and increasingly valuable resource for monitoring and understanding our planet, especially in the context of global change. EO data from the U.S. federal government National Aeronautics and Space Administration (NASA) are among the richest and longest-standing in the field. Iconic and widely NASA EO data collections include the MODIS Land products [@JUSTICE20023], the VIIRS products - which continue the legacy of MODIS [@ROMAN2024113963], and the GPM mission products [@SkofronickJackson2017]. Collectively, these products have provided essential data for over 20 years, enabling the study of Earth's dynamics, including global land cover, vegetation health, burned areas, land surface temperature, rainfall patterns. They support research in climate change, disaster response, biodiversity, ecosystem monitoring, ecology, public health, and more [@modis-applications].
+Data from Earth Observation satellites are a crucial and increasingly valuable resource for monitoring and understanding our planet, especially in the context of global change. EO data from the NASA are among the richest and longest-standing in the field. Iconic and widely NASA EO data collections include the MODIS Land products [@JUSTICE20023], the VIIRS products - which continue the legacy of MODIS [@ROMAN2024113963], and the GPM mission products [@SkofronickJackson2017]. Collectively, these products have provided essential data for over 20 years, enabling the study of Earth's dynamics, including global land cover, vegetation health, burned areas, land surface temperature, rainfall patterns. They support research in climate change, disaster response, biodiversity, ecosystem monitoring, ecology, public health, and more [@modis-applications].
 
-Despite the increasing availability and utility of these EO data, accessing and using them presents several challenges [@AGNOLI2023122357]. These data are typically distributed as multidimensional layers over vast areas. This results in large file sizes, making accessibility and processing difficult, especially when large time series are needed. This problem is particularly acute in developing regions where internet infrastructure, needed to access the data, can be limited. The complexity of accessing EO data can result in obstacles, such as underutilization or the creation of siloed data processing workflows, which separate data extraction from pre-processing and analysis, thereby hindering transparent and reproducible open science practices. While some powerful tools, such as [Google Earth Engine](https://earthengine.google.com/) [@GORELICK201718] offer some solutions to these problems, they also have important limitations, such as proprietary and energy-intensive software. Altogether, these barriers hinder the full potential of Earth observation data to support global research and decision-making. Efforts to simplify, lighten and streamline access to these data, while maintaining an open-source and open-science framework, are essential to overcoming these obstacles and maximizing the benefits of satellite data for all.
+Despite the growing availability and usefulness of these Earth Observation (EO) data, accessing and utilizing it poses significant challenges [@AGNOLI2023122357]. These data are often distributed as large, multidimensional layers, making them difficult to access and process, particularly for extended time series and in regions with limited internet infrastructure. This complexity can lead to underutilization or fragmented data processing workflows, which hinders transparent and reproducible open science. While powerful tools like [Google Earth Engine](https://earthengine.google.com/) [@GORELICK201718] offer some solutions, they come with limitations such as proprietary and energy-intensive software. To fully realize the potential of EO data for global research and decision-making, it is crucial to simplify, lighten and streamline access while maintaining an open-source framework.
 
[…]
