### 2024-08-26 4ada51f
update readme and joss paper
