+++ README.md
@@ -70,6 +70,12 @@ interest is only 30 km x 30 km wide and you need 2 bands ! Moreover,
 data **over long time series** and **over areas**, rather than short
 time series and points.
 
+Importantly, the robust, sustainable, and cost-free [foundational
+framework](#foundational-framework) of `modisfast`, both for the data
+provider (NASA) and the software (R, OPeNDAP, the `tidyverse` and `GDAL`
+suite of packages and software), guarantees the long-term reliability
+and open-source nature of the package.
+
 Below is a comparison of modisfast with other packages available for
 downloading chunks of MODIS or VIIRS data :
 
@@ -1862,13 +1868,24 @@ Rainfall
 
 ## Foundational framework
 
-`modisfast` is an R wrapper for OPeNDAP (*Open-source Project for a
-Network Data Access Protocol*). When utilized by data providers, such as
-those managing many NASA datasets, OPeNDAP allows for subsetting
-portions of Earth observation cubes based on any dimension by specifying
-filters in a URL. `modisfast` facilitates this process by constructing
-the URL based on the spatial, temporal, and dimensional filters provided
-by the user in the function `mf_get_url()`.
+Technically, `modisfast` is a programmatic interface (R wrapper) to
+several NASA OPeNDAP (<https://www.opendap.org/>) servers. OPeNDAP is
+the acronym for *Open-source Project for a Network Data Access Protocol*
+and designates both the software, the access protocol, and the
+corporation that develops them. The OPeNDAP is designed to simplify
+access to structured and high-volume data, such as satellite products,
+over the Web. It is a collaborative effort involving multiple
+institutions and companies, with open-source code, free software, and
+adherence to the [Open Geospatial Consortium](https://www.ogc.org/)
+(OGC) standards. It is widely used by NASA, which partly finances it.
+
+A key feature of OPeNDAP is its capability to apply filters at the data
+download process, ensuring that only the necessary data is retrieved.
+These filters, specified within a URL, can be spatial, temporal, or
+dimensional. Although powerful, OPeNDAP URLs are not trivial to build.
+`modisfast` facilitates this process by constructing the URL based on
+the spatial, temporal, and dimensional filters provided by the user in
+the function `mf_get_url()`.
 
 Let’s take an example to understand.
 
