+++ joss-paper/paper.bib
@@ -72,48 +72,6 @@ abstract = {This paper provides a review and summary status of the research unde
   pages = {1679–1695}
 }
 
-  @Manual{terra,
-    title = {terra: Spatial Data Analysis},
-    author = {Robert J. Hijmans},
-    year = {2024},
-    note = {R package version 1.7-78},
-    url = {https://CRAN.R-project.org/package=terra},
-    doi = {10.32614/CRAN.package.terra}
-  }
-
-
-  @Article{tidyverse,
-    title = {Welcome to the {tidyverse}},
-    author = {Hadley Wickham and Mara Averick and Jennifer Bryan and Winston Chang and Lucy D'Agostino McGowan and Romain François and Garrett Grolemund and Alex Hayes and Lionel Henry and Jim Hester and Max Kuhn and Thomas Lin Pedersen and Evan Miller and Stephan Milton Bache and Kirill Müller and Jeroen Ooms and David Robinson and Dana Paige Seidel and Vitalie Spinu and Kohske Takahashi and Davis Vaughan and Claus Wilke and Kara Woo and Hiroaki Yutani},
-    year = {2019},
-    journal = {Journal of Open Source Software},
-    volume = {4},
-    number = {43},
-    pages = {1686},
-    doi = {10.21105/joss.01686},
-  }
-
-    @Manual{gdal,
-    title = {{GDAL/OGR} Geospatial Data Abstraction software Library},
-    author = {{GDAL/OGR contributors}},
-    organization = {Open Source Geospatial Foundation},
-    year = {2024},
-    url = {https://gdal.org},
-    doi = {10.5281/zenodo.5884351},
-  }
-
-  @Article{sf,
-  author = {Edzer Pebesma},
-  title = {{Simple Features for R: Standardized Support for Spatial Vector Data}},
-  year = {2018},
-  journal = {{The R Journal}},
-  doi = {10.32614/RJ-2018-009},
-  url = {https://doi.org/10.32614/RJ-2018-009},
-  pages = {439--446},
-  volume = {10},
-  number = {1},
-}
-
 
 @Manual{modistools,
          title = "The MODISTools package: an interface to the MODIS Land Products Subsets Web Services",
@@ -157,19 +115,4 @@ abstract = {This paper provides a review and summary status of the research unde
   version      = {v.1.0.0-alpha},
   doi          = {10.5281/zenodo.3945409},
   url          = {https://doi.org/10.5281/zenodo.3945409}
-}
-
-@article{GORELICK201718,
-title = {Google Earth Engine: Planetary-scale geospatial analysis for everyone},
-journal = {Remote Sensing of Environment},
-volume = {202},
-pages = {18-27},
-year = {2017},
-note = {Big Remotely Sensed Data: tools, applications and experiences},
-issn = {0034-4257},
-doi = {10.1016/j.rse.2017.06.031},
-url = {https://www.sciencedirect.com/science/article/pii/S0034425717302900},
-author = {Noel Gorelick and Matt Hancher and Mike Dixon and Simon Ilyushchenko and David Thau and Rebecca Moore},
-keywords = {Cloud computing, Big data, Analysis, Platform, Data democratization, Earth Engine},
-abstract = {Google Earth Engine is a cloud-based platform for planetary-scale geospatial analysis that brings Google's massive computational capabilities to bear on a variety of high-impact societal issues including deforestation, drought, disaster, disease, food security, water management, climate monitoring and environmental protection. It is unique in the field as an integrated platform designed to empower not only traditional remote sensing scientists, but also a much wider audience that lacks the technical capacity needed to utilize traditional supercomputers or large-scale commodity cloud computing resources.}
 }
