+++ README.md
@@ -7,7 +7,7 @@
 <!-- badges: start -->
 
 [![licence](https://img.shields.io/badge/Licence-GPL--3-blue.svg)](https://www.r-project.org/Licenses/GPL-3)
-[![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/modisfast)](https://cran.r-project.org/package=modisfast)
+[![CRAN_Status_Badge](https://www.r-pkg.org/badges/version/modisfast)](https://cran.r-project.org/package=modisfast)
 [![Github_Status_Badge](https://img.shields.io/badge/Github-0.9.2-blue.svg)](https://github.com/ptaconet/modisfast)
 [![SWH](https://archive.softwareheritage.org/badge/origin/https://github.com/ptaconet/modisfast/)](https://archive.softwareheritage.org/browse/origin/?origin_url=https://github.com/ptaconet/modisfast)
 [![R-CMD-check](https://github.com/ptaconet/modisfast/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/ptaconet/modisfast/actions/workflows/R-CMD-check.yaml)
@@ -40,8 +40,8 @@ This package enables to build OPeNDAP (https) URLs given input parameters such a
 available</a><br> • <a href="#foundational-framework">Foundational
 framework </a><br> •
 <a href="#comparison-with-similar-r-packages">Comparison with similar R
-packages</a><br> • <a href="#future-developments">Future
-developments</a><br> • <a href="#citation">Citation</a><br> •
+packages</a><br> • <a href="#citation">Citation</a><br> •
+<a href="#future-developments">Future developments</a><br> •
 <a href="#contributing">Contributing</a><br> •
 <a href="#acknowledgments">Acknowledgments</a><br>
 </p>
@@ -49,24 +49,23 @@ developments</a><br> • <a href="#citation">Citation</a><br> •
 ## Overview
 
 **`modisfast`** is an R package designed for **easy** and **fast**
-downloads and import of some widely-used satellite-derived environmental
-data, including
+downloads of
 [**MODIS**](https://lpdaac.usgs.gov/data/get-started-data/collection-overview/missions/modis-overview/)
 Land products,
 [**VIIRS**](https://lpdaac.usgs.gov/data/get-started-data/collection-overview/missions/s-npp-nasa-viirs-overview/)
-Land products, and [**GPM**](https://gpm.nasa.gov/data/) products
-(Global Precipitation Measurement Mission).
+Land products, and [**GPM**](https://gpm.nasa.gov/data/) (Global
+Precipitation Measurement Mission) Earth Observation data.
 
-**`modisfast`** uses the abilities offered by the
+`modisfast` uses the abilities offered by the
 [OPeNDAP](https://www.opendap.org/about/) framework (*Open-source
 Project for a Network Data Access Protocol*) to download a subset of
-Earth science data cube, along spatial, temporal or any other data
+Earth Observation data cube, along spatial, temporal or any other data
 dimension (depth, …). This way, it reduces downloading time and disk
 usage to their minimum : no more 1° x 1° MODIS tiles with 10 bands when
 your region of interest is only 30 km x 30 km wide and you need 2 bands
[…]
