### 2024-10-10 #7343 @editorialbot
[REVIEW]: modisfast: An R package for fast and efficient access to MODIS, VIIRS and GPM Earth Observation data
**Submitting author:** @ptaconet (<a href="http://orcid.org/0000-0001-7429-7204">Paul Taconet</a>)
**Repository:** https://github.com/ptaconet/modisfast
**Branch with paper.md** (empty if default branch):
**Version:** 0.9.2
