### 2024-10-10 #7343 @mikemahoney218
👋🏼 @ptaconet, @rmendels, @V-Agudetse this is the review thread for the paper. Just about all of our communications will happen here from now on 😄

```@editorialbot generate my checklist```

We aim for reviews to be completed within about 4 weeks. Please let me know if you require some more time.

Please feel free to ping me (@mikemahoney218) if you have any questions/concerns!
