### 2024-10-17 c92980b
* used the cli package for comments (when verbose = TRUE) * solving issue with tile merging at extreme latitude * adding a fileSizeEstimated column in output of mf_get_url
