### 2024-10-21 #7343 @rmendels
I am not certain if this is how we add comments beside the checklist.  I ran the present repository against devtools::check(),  latest version,  had  one minor quibble with where the Citation was.

Also ran the repository through the package 'goodpractice',  I suggest the author  should too.  Clearly a lot of the items are individual taste.  There were:

1.  A lot of warnings about line length.  Not a must but something the author might look at.
2.  A lot of warnings about wanting a test for every single function.  I have not looked yet to see if there is an adequate level of testing,  but again something the author might look at.
3.  A number of other suggestions about safe coding that perhaps should be looked at.  There were really only a few,  but they occurred in multiple places,  so it generates a number of warnings.
