### 2024-10-21 #7343 @rmendels
I should add that devtools::check() fails on the tests, if I am reading the output correctly it is because accessing at least some of the sites requires credentials,  and of course those can not actually be put into public facing tests,  but I also do not see comments or instructions on how to run the tests with credentials.
