### 2024-10-21 #7343 @mikemahoney218
Thank you, @rmendels! Would you be willing to post these as issues on the modisfast repository (and mention openjournals/joss-reviews#7343 in the issue body so that it links back here)?
