### 2024-10-23 ptaconet/modisfast#8 @V-Agudetse
[JOSS Review] Community guidelines
Hello @ptaconet,

I am opening this issue in the context of the review of the modisfast publication in JOSS (openjournals/joss-reviews#7343) (really nice package with great usage examples, btw :) ).

I have a couple of notes related to the community guidelines after reviewing [CONTRIBUTING.md](https://github.com/ptaconet/modisfast/blob/master/CONTRIBUTING.md):

Blocking:
- Some of the links in the document (as well as in other documents, such as README.md) that point to files within the modisfast repository are broken because they point to a non-existing "main" branch instead of the current master branch. This interferes with the clarity of the guidelines so the links should be fixed before the "community guidelines" box can be checked.

Non-blocking:
- This is not part of the JOSS review requirements, but the contribution guidelines state that code contributions should follow the Tidyverse style guide, but the style of a large part of the code does not follow Tidyverse conventions, especially in regards to line length and syntax issues such as spacing. I can recommend the styler package as a tool to quickly format the code to be in accordance with Tidyverse conventions.

Cheers,

Victòria A.
