### 2024-10-23 ptaconet/modisfast#9 @rmendels
Citations in Paper
Please refer to this [JOSS review issue](https://github.com/openjournals/joss-reviews/issues/7343)

R packages on CRAN have a formal citation even if there isn't a paper,  as does the OPeNDAP protocol.  It is good that you link to these,  but the formal citations should also be in the references.  This does not count in the paper word count limit.
