### 2024-10-23 ptaconet/modisfast#10 @rmendels
Code Cleanup and Documentation
Please refer to this [JOSS review issue](https://github.com/openjournals/joss-reviews/issues/7343).

 I ran the present repository against devtools::check(), latest version, had one minor quibble with where the Citation was.  Also the tests fail when run this way,  because userid and password are required to access some of the data and of course you can't put that in a public repository.  But I did not see anything that would warn a user of this,  as well as instructions on how to run the tests since they will fail if trying to run automatically (note this is an item in the review checklist)  These could be comments in the test code.

Also ran the repository through the package 'goodpractice', I suggest the you should too. Clearly a lot of the items are individual taste. There were:

A lot of warnings about line length. Not a must but something the you might look at.
A lot of warnings about wanting a test for every single function. I have not looked yet to see if there is an adequate level of testing, but again something the you might look at.
A number of other suggestions about safe coding that perhaps should be looked at. There were really only a few, but they occurred in multiple places, so it generates a number of warnings.
