+++ joss-paper/paper.bib
@@ -116,3 +116,101 @@ abstract = {This paper provides a review and summary status of the research unde
   doi          = {10.5281/zenodo.3945409},
   url          = {https://doi.org/10.5281/zenodo.3945409}
 }
+
+  @Misc{modisfast,
+    note = {R package version 1.0.0},
+    title = {modisfast : fast and efficient access to MODIS data with R},
+    author = {Paul Taconet and Nicolas Moiroux},
+    url = {https://hal.science/hal-04603903},
+    year = {2024},
+    month = {Jun},
+    swhid =
+      {swh:1:dir:21b5ddcecb39e683c9f2fc5c135f23dc1b36fe28;origin=https://github.com/ptaconet/modisfast;visit=swh:1:snp:4064f0aae2b29808e9cbd096afaa495fd1360f78;anchor=swh:1:rev:9e5b2a456a0e1acd8ab34e6909424ae6c403150d},
+    repository = {git+https://github.com/ptaconet/modisfast.git},
+    license = {https://spdx.org/licenses/GPL-3.0+},
+    keywords = {MODIS ; Earth datacube ; R ; OPeNDAP ; VIIRS ; Earth
+      science},
+    hal_id = {hal-04603903},
+    hal_version = {v1},
+  }
+
+@software{opendap1,
+  author       = {James Gallagher and
+                  Nathan Potter and
+                  rmorris2342 and
+                  kyang2014 and
+                  Captain James Tiberius Kirk and
+                  Kodi Neumiller and
+                  The Robot Travis and
+                  tsgouros and
+                  blackone-sudo and
+                  Slav Korolev and
+                  Dan Horák and
+                  Ethan Davis and
+                  H. Joe Lee and
+                  yuanho and
+                  Bo Anderson and
+                  Dan Holloway and
+                  Orion Poplawski and
+                  Ryan Carsten Schmidt and
+                  Sam Lloyd},
+  title        = {OPENDAP/libdap4: Hyrax-1.17.0},
+  month        = jan,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {3.21.0-27},
+  doi          = {10.5281/zenodo.10564122},
+  url          = {https://doi.org/10.5281/zenodo.10564122}
+}
+
+@software{opendap2,
+  author       = {James Gallagher and
+                  Nathan Potter and
+                  kyang2014 and
+                  The Robot Travis and
+                  Kodi Neumiller and
+                  Slav Korolev and
+                  Corey Hemphill and
+                  H. Joe Lee and
+                  Riley Rimer and
+                  Sam Lloyd and
+                  blackone-sudo and
+                  Uday Kari and
+                  Dan Holloway and
+                  rmorris2342 and
+                  Orion Poplawski and
+                  Captain James Tiberius Kirk and
+                  Edward Hartnett and
+                  Ethan Davis and
+                  Lewis John McGibbney and
+                  tsgouros and
+                  ptype},
+  title        = {OPENDAP/bes: BES-3.21.0-46 for Hyrax-1.17.0},
+  month        = jan,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {3.21.0-46},
+  doi          = {10.5281/zenodo.10564745},
+  url          = {https://doi.org/10.5281/zenodo.10564745}
+}
+
+@software{opendap3,
+  author       = {Nathan Potter and
+                  The Robot Travis and
+                  James Gallagher and
+                  Uday Kari and
+                  Slav Korolev and
+                  Kodi Neumiller and
+                  ideaesb and
+                  Riley Rimer and
+                  Dan Holloway and
+                  Sam Lloyd},
[…]
