### 2024-10-24 25c842b
Fix first part of issue #10 (citation location)
