### 2024-10-24 ptaconet/modisfast#10 @ptaconet
Commit https://github.com/ptaconet/modisfast/commit/25c842b12e836fc358d42e00f2510faeaa2b8ff1  fixes the Citation problem
