### 2024-10-24 ptaconet/modisfast#8 @ptaconet
Hi V-Agudetse, thanks for your feedback :)
Commit https://github.com/ptaconet/modisfast/commit/8c438631c41d5270036efde385ec9bc78d1035ec fixes the "Blocking" part of the issue.
I am now working on the Non-blocking part
