### 2024-10-24 b15397d
fix part of "Non-blocking" issue #8 and part of issue #10
