+++ R/mf_list_variables.R
@@ -15,85 +15,88 @@
 #' @importFrom stringr str_match word
 #' @import purrr dplyr httr
 #' @examples
-#'
 #' \dontrun{
 #' # login to Earthdata
-#' log <- mf_login(c("earthdata_un","earthdata_pw"))
+#' log <- mf_login(c("earthdata_un", "earthdata_pw"))
 #'
 #' # Get the variables available for the collection MOD11A1.061
 #' (df_varinfo <- mf_list_variables("MOD11A1.061"))
-#'}
+#' }
 #'
-
-mf_list_variables<-function(collection,credentials=NULL){  # for a given collection, get the available variables and associated information
+mf_list_variables <- function(collection, credentials = NULL) { # for a given collection, get the available variables and associated information
 
   .testIfCollExists(collection)
   .testInternetConnection()
   .testLogin(credentials)
 
-  httr::set_config(httr::authenticate(user=getOption("earthdata_user"), password=getOption("earthdata_pass"), type = "basic"))
+  httr::set_config(httr::authenticate(user = getOption("earthdata_user"), password = getOption("earthdata_pass"), type = "basic"))
 
-  opendapMetadata <- opendapMetadata_internal[which(opendapMetadata_internal$collection==collection),]
+  opendapMetadata <- opendapMetadata_internal[which(opendapMetadata_internal$collection == collection), ]
 
-  URL<-opendapMetadata$url_opendapexample
+  URL <- opendapMetadata$url_opendapexample
 
-  InfoURL<-paste0(URL,".info")
-  vector_response<-httr::GET(InfoURL)
-  vector_response<-httr::GET(vector_response$url)
-  if(vector_response$status_code==400){ stop("Bad request\n")}
-  vector_content<-httr::content(vector_response,"text",encoding="UTF-8")
-  vector_html<-xml2::read_html(vector_content)
-  tab<-rvest::html_table(vector_html)
-  if(purrr::is_empty(tab)){ stop("The server might be temporarily unavailable. Try again later. Paste ",InfoURL," to check the error message in your brower\n")}
-  tab<-tab[[length(tab)]]
-  colnames(tab)<-c("name","all_info")
-  tab$name<-gsub(":","",tab$name)
-  tab$long_name<-stringr::str_match(tab$all_info, "long_name: (.*?)\n")[,2]
-  tab$units<-stringr::str_match(tab$all_info, "units: (.*?)\n")[,2]
+  InfoURL <- paste0(URL, ".info")
+  vector_response <- httr::GET(InfoURL)
+  vector_response <- httr::GET(vector_response$url)
+  if (vector_response$status_code == 400) {
+    stop("Bad request\n")
+  }
+  vector_content <- httr::content(vector_response, "text", encoding = "UTF-8")
+  vector_html <- xml2::read_html(vector_content)
+  tab <- rvest::html_table(vector_html)
+  if (purrr::is_empty(tab)) {
+    stop("The server might be temporarily unavailable. Try again later. Paste ", InfoURL, " to check the error message in your brower\n")
+  }
+  tab <- tab[[length(tab)]]
+  colnames(tab) <- c("name", "all_info")
+  tab$name <- gsub(":", "", tab$name)
+  tab$long_name <- stringr::str_match(tab$all_info, "long_name: (.*?)\n")[, 2]
+  tab$units <- stringr::str_match(tab$all_info, "units: (.*?)\n")[, 2]
 
-  DdsURL<-paste0(URL,".dds")
-  vector_response<-httr::GET(DdsURL)
[…]
