+++ R/internal_tests.R
@@ -3,16 +3,16 @@
 #' @noRd
 #'
 
-.testIfCollExists<-function(collection){
-
-  if(!inherits(collection,"character") || length(collection)!=1 ){stop("Argument collection must be a character string of length 1")}
-
-  collection<-opendapMetadata_internal$collection[which(opendapMetadata_internal$collection==collection)]
-
-  if(length(collection)==0){
-    stop("The collection that you specified does not exist. Check mf_list_collections() to see which collections are implemented\n")
+.testIfCollExists <- function(collection) {
+  if (!inherits(collection, "character") || length(collection) != 1) {
+    stop("Argument collection must be a character string of length 1")
   }
 
+  collection <- opendapMetadata_internal$collection[which(opendapMetadata_internal$collection == collection)]
+
+  if (length(collection) == 0) {
+    stop("The collection that you specified does not exist. Check mf_list_collections() to see which collections are implemented\n")
+  }
 }
 
 #' @name .testIfVarExists
@@ -20,71 +20,79 @@
 #' @noRd
 #'
 
-.testIfVarExists<-function(specified_variables,existing_variables){
-
- diff_vars <- NULL
- diff_vars <- setdiff(specified_variables,existing_variables)
- if(length(diff_vars)>0){stop("Specified variables do not exist or are not extractable for the specified collection. Use the function mf_list_variables to check which variables are available and extractable for the collection\n")}
-
+.testIfVarExists <- function(specified_variables, existing_variables) {
+  diff_vars <- NULL
+  diff_vars <- setdiff(specified_variables, existing_variables)
+  if (length(diff_vars) > 0) {
+    stop("Specified variables do not exist or are not extractable for the specified collection. Use the function mf_list_variables to check which variables are available and extractable for the collection\n")
+  }
 }
 
-#.testIfVarExists2<-function(collection,specified_variables,credentials=NULL){
+# .testIfVarExists2<-function(collection,specified_variables,credentials=NULL){
 #  variables <- NULL
 #  .testIfCollExists(collection)
 #  .testLogin(credentials)
 #  variables <- mf_list_variables(collection,credentials)
 #  variables <- variables$name
 #  .testIfVarExists2(variables,specified_variables)
-#}
+# }
 
 #' @name .testLogin
 #' @title Test login, else log
 #' @noRd
 
-.testLogin<-function(credentials=NULL){
-
+.testLogin <- function(credentials = NULL) {
   mf_login <- NULL
 
-  if(!is.null(credentials) || is.null(getOption("earthdata_mf_login"))){
-    mf_login <- mf_login(credentials,source = "earthdata")
+  if (!is.null(credentials) || is.null(getOption("earthdata_mf_login"))) {
+    mf_login <- mf_login(credentials, source = "earthdata")
     return(mf_login)
   }
-
 }
 
 #' @name .testRoi
 #' @title Test roi
 #' @noRd
 
-.testRoi<-function(roi){
-  #if(!inherits(roi,"sf") || as.character(unique(sf::st_geometry_type(roi)))!="POLYGON" || !("id" %in% colnames(roi)) || !("geom" %in% colnames(roi)) || length(which(is.na(roi$id))) || length(which(is.na(roi$geom)))){stop("Argument roi must be an object of class sf or sfc with POLYGON-type feature geometry and at least two columns : 'id' and 'geom' that must not be NULL or NA")}
[…]
