+++ R/internal_ancillary.R
@@ -21,27 +21,26 @@
 
 
 
-.getMODIStileNames<-function(roi,type){
-
+.getMODIStileNames <- function(roi, type) {
   . <- Name <- NULL
 
   .testRoi(roi)
 
   roi_id <- roi$id
 
-  #modis_tile = sf::read_sf("https://modis.ornl.gov/files/modis_sin.kmz") %>%
-  if(type=="modis"){
+  # modis_tile = sf::read_sf("https://modis.ornl.gov/files/modis_sin.kmz") %>%
+  if (type == "modis") {
     tiling_system <- modis_tiles
-  } else if (type=="suomi"){
+  } else if (type == "suomi") {
     tiling_system <- suomi_tiles
   }
 
-  sf::st_agr(tiling_system) = "constant"
-  sf::st_agr(roi) = "constant"
+  sf::st_agr(tiling_system) <- "constant"
+  sf::st_agr(roi) <- "constant"
 
-   roi <- sf::st_transform(roi,4326) %>%
-     sf::st_bbox() %>%
-     sf::st_as_sfc()
+  roi <- sf::st_transform(roi, 4326) %>%
+    sf::st_bbox() %>%
+    sf::st_as_sfc()
 
   modis_tile <- tiling_system %>%
     sf::st_intersection(roi) %>%
@@ -51,21 +50,21 @@
 
   modis_tile <- modis_tile$Name
 
-    all_modis_tiles <- NULL
-    for(i in 1:length(modis_tile)){
-    th_modis_tile<-modis_tile[i] %>%
+  all_modis_tiles <- NULL
+  for (i in seq_along(modis_tile)) {
+    th_modis_tile <- modis_tile[i] %>%
       unique() %>%
-      stringr::str_replace_all(c(" "="",":"=""))
-    for (i in 1:9){
-      th_modis_tile<-gsub(paste0("h",i,"v"),paste0("h0",i,"v"),th_modis_tile)
+      stringr::str_replace_all(c(" " = "", ":" = ""))
+    for (i in 1:9) {
+      th_modis_tile <- gsub(paste0("h", i, "v"), paste0("h0", i, "v"), th_modis_tile)
     }
-    if(nchar(th_modis_tile)!=6){
-      th_modis_tile<-paste0(substr(th_modis_tile,1,4),"0",substr(th_modis_tile,5,5))
+    if (nchar(th_modis_tile) != 6) {
+      th_modis_tile <- paste0(substr(th_modis_tile, 1, 4), "0", substr(th_modis_tile, 5, 5))
     }
-    all_modis_tiles<-c(all_modis_tiles,th_modis_tile)
-    }
+    all_modis_tiles <- c(all_modis_tiles, th_modis_tile)
+  }
 
-    roi_id <- rep(roi_id,length(all_modis_tiles))
+  roi_id <- rep(roi_id, length(all_modis_tiles))
 
   return(list(all_modis_tiles = all_modis_tiles, roi_id = roi_id))
 }

[…]
