### 2024-10-24 ptaconet/modisfast#10 @ptaconet
Commit https://github.com/ptaconet/modisfast/commit/b15397d275a813abf77e8702c03927996d8fe159 fixes part of the coding style suggestions. As suggested, I have ran the repository through the package 'goodpractice' (I did not know this package, thanks for the mention!). I have made the requested changes, except for the line length (I must say it is very difficult to respect the 80 characters line length...). For the functions testing, I am currently working on writing tests for the most important ones.
