### 2024-10-24 ptaconet/modisfast#8 @ptaconet
Commit https://github.com/ptaconet/modisfast/commit/b15397d275a813abf77e8702c03927996d8fe159 fixes part of Non-blocking suggestions. As suggested, I have ran the repository through the package 'styler' (I did not know this package, thanks for the mention!). I have made the requested changes, except for the line length limit (I must say it is very difficult to respect the 80 characters line length...). In order to take into account this limitation, in the Contributing.md file I have changed the sentence "We use the Tidyverse style guide for writing R code" to "We use as much as possible the Tidyverse style guide for writing R code - except for line length".
