### 2024-10-24 ptaconet/modisfast#8 @ptaconet
This issue is now **a priori** solved, @V-Agudetse feel free to give your feedback so that I can close it.
