### 2024-10-24 ptaconet/modisfast#8 @V-Agudetse
Everything looks good @ptaconet; from my side this can be closed!
