### 2024-10-24 ptaconet/modisfast#9 @ptaconet
@rmendels please indicate if everything is good for you (or if you have a suggestion for a better citation of OPeNDAP) so that I can close the issue :)
