### 2024-10-24 ptaconet/modisfast#9 @rmendels
@ptaconet -  I actually emailed OPenDAP support the other day about their preferred citation and have not received a reply which has happened a number of times recently.  I am good with what you have.  Thanks.
