### 2024-10-30 e683529
adding tests of the functionality of the package and instructions on how to run them (issue #10 and #11)
