### 2024-10-30 d77f8b4
adding the high-level function 'modisfast()' which enables to execute the whole workflow at once (login, get URL, download data, import data)
