### 2024-10-30 110ae2a
lights updates on the documentations of the functions + adding the cli package for messages when verbose
