### 2024-10-30 ptaconet/modisfast#10 @ptaconet
Commit https://github.com/ptaconet/modisfast/commit/e683529ba0e6d508c4d5f597d3f2933855b17bb6 adds tests for the main functions of the package, and documentation on how to execute them, as outlined in the Readme - see : https://github.com/ptaconet/modisfast/tree/master?tab=readme-ov-file#manual-testing-of-the-functionality

In addition, devtools::check() now does not fail anymore when running the package against devtools::check()
