### 2024-10-30 #7343 @ptaconet
Hi @rmendels and @V-Agudetse ! Thank you for the first steps of your review :)

Just a quick message to warn you that in order to answer the "Automated tests" point of the checklist, I have added additional tests for the main functions of the package, as shown in commit https://github.com/ptaconet/modisfast/commit/e683529ba0e6d508c4d5f597d3f2933855b17bb6, and I have outlined how to execute them in the documentation : https://github.com/ptaconet/modisfast/tree/master?tab=readme-ov-file#manual-testing-of-the-functionality
