### 2024-10-30 ptaconet/modisfast#12 @rmendels
R help problem
Please refer to this https://github.com/openjournals/joss-reviews/issues/7343.

I have updated to your latest on github just a few minutes ago, so that you know what version I am referring to.  I am running the latest R and RStudio.  When I go "?modisfast"  a proper page comes up,  but if I then click on "[Package modisfast version 0.9.1 [Index](http://127.0.0.1:18327/help/00Index.html)]"    I get "URL '/help/00Index.html' not found" .  The appropriate function help do work,  ie "?mf_download_data"
