### 2024-10-30 ptaconet/modisfast#12 @ptaconet
you mean, when you click in what I point out with the red arrow in the attached image ?
![Capture d’écran du 2024-10-30 18-57-45](https://github.com/user-attachments/assets/25ba59e9-19a3-476b-b42b-f7bd7e6685e5)
