### 2024-10-30 ptaconet/modisfast#12 @ptaconet
I am not getting this problem, even when trying to install the development version of the package on another machine..
Maybe restarting your R session may solve the problem ? see https://stackoverflow.com/questions/67247051/url-help-library-package-r-html-00index-html-not-found-when-using-devtools
