### 2024-10-30 ptaconet/modisfast#12 @rmendels
@ptaconet   Yes that is what I mean.  I will try on another computer in a bit,  and will let you know.
