### 2024-10-30 ptaconet/modisfast#13 @rmendels
Comment/problems with examples and vignettes
Please refer to this https://github.com/openjournals/joss-reviews/issues/7343.

Summary:

1.  All examples for documents routines run fine.
2. In vignettes,  all MODIS examples work,  all GPM examples fail, go into endless loop trying to download
3. In vignettes, put all needed libraries at top,  the one most often not is "mapview"
4. package ""appeears" is referenced but is not in DESCRIPTION, probably need it as a suggest.
5.  run all vignettes through a spell-checker,  several typos.

More specifics:

1.  For get_started.RMD

a.  For new users the 3 options to get login credentials will be confusing,  they will not know which to use.  More is needed on which to use when

b. "unlink(target_fold, recursive = T)" can throw an error.  Check for existence before unlinking
c. by default,  "target_folder" is in user space.  A new user may not catch that,  it is generally frowned on,  default should be to temporary space with instructions to user to edit for their own space.
d. The comments beginning " <!-- The collections generally contain several variables  " is really confusing.  I am a long time user of OPeNDAP,  and I found it confusing.  A more thought out description of the file structure and how that relates to the various functions is needed.
e.verbose mode should do more,  should allow me to see the commands that are being sent,  I believe you are using httr so it is sy to implement.
f.  what appears like an error when first accessing GPM data:

```r
  > urls_gpm <- mf_get_url(
    +     collection = "GPM_3IMERGM.07",
    +     variables = c("precipitation"),  # get the available variables with : mf_list_variables("GPM_3IMERGM.07")
    +     roi = roi,
    +     time_range = time_range )
  Building the URLs...
  OK
  Warning messages:
    1: In .getVarVector(OpendapURL, odap_coll_info$dim_lon) :
    NAs introduced by coercion
  2: In .getVarVector(OpendapURL, odap_coll_info$dim_lat) :
    NAs introduced by coercion
  ```
2. For perf_comp.Rmd

a.  The line:
```r
token <- rs_login(user = Sys.getenv("earthdata_un"))
```

will cause problems.  More is needed for a user to run successfully

b.  For package "MODIS" no need to include anything besides description at beginning since it is not really available

Nice package,  hope this helps.
