### 2024-10-30 ptaconet/modisfast#12 @rmendels
@ptaconet tried on a new computer, it worked.  I am closing this
