### 2024-10-30 ptaconet/modisfast#13 @ptaconet
Thank you so much @rmendels for your thourough testing of the package. I will work on solving all these issues very soon.
