### 2024-10-31 bd364ba
fix issue #13 points 2a and 2b
