### 2024-11-01 ptaconet/modisfast#10 @ptaconet
Thanks a lot @V-Agudetse for both the bugs reports and suggestions, I will work on this very soon !
