### 2024-11-04 ptaconet/modisfast#10 @ptaconet
closing this issue as it was transferred to issue #15
