+++ tests/testthat.R
@@ -3,10 +3,10 @@ library(modisfast)
 
 ### Instructions to run the tests in order to verify the functionality of modisfast :
 
-# - open the file test/testthat/test-modisfast.R
-# - uncomment the 2 first two lines of the script (earthdata_un = "my_username" and earthdata_pw = "my_password")
+# - open the file : test/testthat/helper-modisfast.R
+# - uncomment the lines : earthdata_un = "my_username" and earthdata_pw = "my_password"
 # - replace "my_username" and "my_password" with your own Earthdata username and password
-# - run devtools::test()  (or run the script test/testthat/test-modisfast.R manually)
+# - run devtools::test()
 
 message("tests are not executed automatically")
 
