+++ tests/testthat/test-modisfast.R
@@ -1,236 +1,0 @@
-
-# earthdata_un = "my_username"
-# earthdata_pw = "my_password"
-
-################
-# test mf_login()
-################
-
-test_that("function mf_login() sends back the expected and outputs and errors when necessary", {
-  skip_on_cran()
-  skip_on_ci()
-  expect_error(mf_login(c(earthdata_un,earthdata_pw,"fd")),"credentials must be a vector character string of length 2 (username and password)\n", fixed = TRUE)
-  expect_error(mf_login(c(earthdata_un)),"credentials must be a vector character string of length 2 (username and password)\n", fixed = TRUE)
-  expect_error(mf_login(c(earthdata_un,"wrong password")))  # error if wrong password is provided
-  expect_error(mf_login(c("wrong username",earthdata_pw)))  # error if wrong username is provided
-  expect_null(mf_login(c(earthdata_un,earthdata_pw)))   # no error if right password or username are provided
-
-})
-
-
-################
-# test mf_list_collections()
-################
-
-test_that("function mf_list_collections() sends back the expected output", {
-  skip_on_cran()
-  skip_on_ci()
-  coll <- mf_list_collections()
-  expect_is(coll,'data.frame')  # output is a data.frame
-})
-
-################
-# test mf_list_variables()
-################
-
-log <- mf_login(c(earthdata_un,earthdata_pw))
-
-test_that("function mf_list_variables() sends back the expected output", {
-
-  skip_on_cran()
-  skip_on_ci()
-
-  # here we test with the collections "MOD13A3.061" and "GPM_3IMERGDF.07"
-  vars_mod13a3 <- mf_list_variables("MOD13A3.061")
-  expect_is(vars_mod13a3,'data.frame')  # output is a data.frame
-  expect_named(vars_mod13a3, c("name","long_name","units","indices","all_info","extractable_with_modisfast")) # column names are ok
-  expect_equal(nrow(vars_mod13a3), 17) # there are 17 rows (corresponding to 17 variables/bands for this collection)
-
-  vars_gpm <- mf_list_variables("GPM_3IMERGDF.07")
-  expect_is(vars_gpm,'data.frame')  # output is a data.frame
-  expect_named(vars_gpm, c("name","long_name","units","indices","all_info","extractable_with_modisfast")) # column names are ok
-  expect_equal(nrow(vars_gpm), 14) # there are 14 rows (corresponding to 14 variables/bands for this collection)
-
-})
-
-
-
-################  ################  ################  ################
-## test downloading and importing of a MODIS datacube
-################  ################  ################  ################
-
-library(sf)
-library(terra)
-
-roi <- st_as_sf(data.frame(id = "madagascar", geom = "POLYGON((41.95 -11.37,51.26 -11.37,51.26 -26.17,41.95 -26.17,41.95 -11.37))"), wkt = "geom", crs = 4326)
-time_range <- as.Date(c("2023-01-01", "2023-03-31"))
-
-
-test_that("function mf_get_url() sends back the expected output and errors when necessary", {
-  skip_on_cran()
-  skip_on_ci()
-  urls <- mf_get_url(collection = "MOD13A3.061",
-                     variables = c("_1_km_monthly_NDVI"),
-                     roi = roi,
-                     time_range = time_range
-  )
-
-  expect_is(urls, "data.frame") # output is a data.frame
-  expect_named(urls, c("id_roi","time_start","collection","name","url")) # column names are ok
-  expect_equal(ncol(urls), 5) # there are 5 columns
[…]
