### 2024-11-06 30f9a3c
update instructions on how to run tests
