### 2024-11-06 20df93f
skip tests if no credentials (or wrong credentials) are provided #issue 15
