+++ tests/testthat/test-mf_login.R
@@ -4,6 +4,8 @@ test_that("function mf_login() sends back the expected outputs and errors when n
 
   skip_on_cran()
   skip_on_ci()
+  skip_if_creds_not_provided()
+  skip_if_wrong_creds()
 
   expect_error(mf_login(credentials = c(earthdata_un,earthdata_pw,"hello")),"credentials must be a vector character string of length 2 (username and password)\n", fixed = TRUE)
   expect_error(mf_login(credentials = c(earthdata_un)),"credentials must be a vector character string of length 2 (username and password)\n", fixed = TRUE)
