### 2024-11-06 dde3bdd
fix bug in readme
