### 2024-11-07 811dd01
adding package appeears in suggestions
