+++ DESCRIPTION
@@ -40,7 +40,8 @@ Suggests:
     mapview,
     rmarkdown,
     spelling,
-    testthat
+    testthat, 
+    appeears
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.3.1
