### 2024-11-07 ptaconet/modisfast#13 @ptaconet
Hi @rmendels , thank you once again for your comments

commits https://github.com/ptaconet/modisfast/commit/bd364ba8584a92c0fb5b8d09cf607094b7dda334 and https://github.com/ptaconet/modisfast/issues/13#ref-commit-f1f9a26 and https://github.com/ptaconet/modisfast/commit/811dd01f4c23fff6fab50b84e6a3f0ca43db0583 solves them.

Do not hesitate to test and tell me if you are ok with this new version, so that I can either propose new modification or close the issue !
