### 2024-11-07 6ee2591
re-run vignette perf_comp
