### 2024-11-07 8f990aa
update verbose for size of downloaded data
