### 2024-11-07 ptaconet/modisfast#13 @ptaconet
Please note that while I have improved the verbose mode (with three modes "quiet", "inform" and "debug"), I could not do much when download is parallelized..
