### 2024-11-07 7bcc0cd
update readme with JOSS badge
