### 2024-11-08 8fffb9e
changed some verbose comment to cli
