### 2024-11-08 6ddb6aa
updating package with several httr::stop_for_status()
