+++ joss-paper/paper.bib
@@ -117,22 +117,19 @@ abstract = {This paper provides a review and summary status of the research unde
   url          = {https://doi.org/10.5281/zenodo.3945409}
 }
 
-  @Misc{modisfast,
-    note = {R package version 1.0.0},
-    title = {modisfast : fast and efficient access to MODIS data with R},
-    author = {Paul Taconet and Nicolas Moiroux},
-    url = {https://hal.science/hal-04603903},
-    year = {2024},
-    month = {Jun},
-    swhid =
-      {swh:1:dir:21b5ddcecb39e683c9f2fc5c135f23dc1b36fe28;origin=https://github.com/ptaconet/modisfast;visit=swh:1:snp:4064f0aae2b29808e9cbd096afaa495fd1360f78;anchor=swh:1:rev:9e5b2a456a0e1acd8ab34e6909424ae6c403150d},
-    repository = {git+https://github.com/ptaconet/modisfast.git},
-    license = {https://spdx.org/licenses/GPL-3.0+},
-    keywords = {MODIS ; Earth datacube ; R ; OPeNDAP ; VIIRS ; Earth
-      science},
-    hal_id = {hal-04603903},
-    hal_version = {v1},
-  }
+@software{modisfast,
+  author       = {Paul Taconet and
+                  Moiroux, Nicolas},
+  title        = {{modisfast: An R package for fast and efficient
+                   access to MODIS, VIIRS and GPM Earth Observation
+                   data}},
+  month        = jul,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {0.9.2},
+  doi          = {10.5281/zenodo.12772739},
+  url          = {https://doi.org/10.5281/zenodo.12772739}
+}
 
 @software{opendap1,
   author       = {James Gallagher and
