### 2024-11-08 #7343 @ptaconet
I have just updated the Zenodo citation to modisfast in the paper, so I regenerate the paper.
