### 2024-11-08 ptaconet/modisfast#16 @V-Agudetse
[JOSS Review] Citation formatting
Hi @ptaconet, this is my only remaining comment related to https://github.com/openjournals/joss-reviews/issues/7343; it's a very minor issue.

According to the style guide (https://pandoc.org/MANUAL.html) the titles of the .bib citation file should be in sentence case. See this section specifically: : https://pandoc.org/MANUAL.html#capitalization-in-titles

Cheers,

Victòria
