### 2024-11-08 ptaconet/modisfast#16 @ptaconet
Hi @V-Agudetse ,

Thanks for this detail.
In the document https://pandoc.org/MANUAL.html#capitalization-in-titles , it is written :

> If you are using a bibtex or biblatex bibliography, then observe the following rules:
>
>   -  English titles should be in title case.

I am actually using a bibtex bibliography, so I guess the titles should be in title case rather than in lower case ? Or am I loosing some detail here ?
