### 2024-11-08 ptaconet/modisfast#16 @V-Agudetse
Yes, forgive me! I meant title case. Examples here: https://apastyle.apa.org/style-grammar-guidelines/capitalization/title-case
