+++ joss-paper/paper.bib
@@ -1,5 +1,5 @@
 @article{JUSTICE20023,
-title = {An overview of MODIS Land data processing and product status},
+title = {An Overview of MODIS Land Data Processing and Product Status},
 journal = {Remote Sensing of Environment},
 volume = {83},
 number = {1},
@@ -18,11 +18,11 @@ author = {Shao, Yang and Taff, Greg and Lunetta, Ross},
 year = {2011},
 month = {01},
 pages = {31-55},
-title = {A review of selected MODIS algorithms, data products, and applications}
+title = {A Review of Selected MODIS Algorithms, Data Products, and Applications}
 }
 
 @article{AGNOLI2023122357,
-title = {Perspectives on user engagement of satellite Earth observation for water quality management},
+title = {Perspectives on User Engagement of Satellite Earth Observation for Water Quality Management},
 journal = {Technological Forecasting and Social Change},
 volume = {189},
 pages = {122357},
@@ -35,7 +35,7 @@ keywords = {Water quality, Earth observation, Satellite remote sensing, Knowledg
 }
 
 @article{ROMAN2024113963,
-title = {Continuity between NASA MODIS Collection 6.1 and VIIRS Collection 2 land products},
+title = {Continuity Between NASA MODIS Collection 6.1 and VIIRS Collection 2 Land Products},
 journal = {Remote Sensing of Environment},
 volume = {302},
 pages = {113963},
@@ -74,7 +74,7 @@ abstract = {This paper provides a review and summary status of the research unde
 
 
 @Manual{modistools,
-         title = "The MODISTools package: an interface to the MODIS Land Products Subsets Web Services",
+         title = "The MODISTools Package: An Interface to the MODIS Land Products Subsets Web Services",
          author = "Koen Hufkens",
          year = "2022",
          url = "https://github.com/ropensci/MODISTools",
@@ -83,7 +83,7 @@ abstract = {This paper provides a review and summary status of the research unde
 
 @Article{MODIStsp,
   author       = "Lorenzo Busetto and Luigi Ranghetti",
-  title        = "MODIStsp: an R package for preprocessing of MODIS Land Products time series",
+  title        = "MODIStsp: An R Package for Preprocessing of MODIS Land Products Time Series",
   journal      = "Computers & Geosciences",
   year         = "2016",
   volume       = "97",
@@ -94,7 +94,7 @@ abstract = {This paper provides a review and summary status of the research unde
 }
 
 @Misc{appeears,
-         title = "The appeears package: an interface to NASA AppEEARS API endpoints",
+         title = "The appeears Package: An Interface to NASA AppEEARS API Endpoints",
          author = "Koen Hufkens and Elio Campitelli",
          year = "2023",
          doi = "10.5281/zenodo.7938190",
@@ -107,7 +107,7 @@ abstract = {This paper provides a review and summary status of the research unde
                   Lesly Bautista and
                   Roy Yali and
                   Antony Barja},
-  title        = {{rgee: An R package for interacting with Google
+  title        = {{rgee: An R Package for Interacting With Google
                    Earth Engine}},
   month        = jul,
   year         = 2020,
@@ -120,9 +120,9 @@ abstract = {This paper provides a review and summary status of the research unde
 @software{modisfast,
   author       = {Paul Taconet and
                   Moiroux, Nicolas},
-  title        = {{modisfast: An R package for fast and efficient
-                   access to MODIS, VIIRS and GPM Earth Observation
-                   data}},
+  title        = {{modisfast: An R Package for Fast and Efficient
+                   Access to MODIS, VIIRS and GPM Earth Observation
+                   Data}},
   month        = jul,
   year         = 2024,
   publisher    = {Zenodo},
