### 2024-11-08 ptaconet/modisfast#16 @ptaconet
Ok, fixed in commit https://github.com/ptaconet/modisfast/issues/16 !
