+++ vignettes/articles/use_case.Rmd
@@ -10,6 +10,8 @@ vignette: >
 format:
   html:
     code-overflow: scroll
+toc:
+  depth: 2
 ---
 
 ```{r, include = FALSE}
