### 2024-11-11 #7343 @mikemahoney218
Posting this here as confirmation from Roy that he's recommending acceptance of this paper.

![image](https://github.com/user-attachments/assets/d5a84d70-a322-42cb-89cb-c3f0c22e7720)
