### 2024-11-11 ptaconet/modisfast#13 @ptaconet
Closing the issue as @rmendels is fine with the overall modifications (see https://github.com/openjournals/joss-reviews/issues/7343#issuecomment-2468260434)
