### 2024-11-12 41615f0
update display of collections in Readme + very minor modifs
