+++ README.md
@@ -240,56 +240,23 @@ Collection
 Source
 </th>
 <th style="text-align:left;">
+Type
+</th>
+<th style="text-align:left;">
 Name
 </th>
 <th style="text-align:left;">
-Type
+Spatial resolution
+</th>
+<th style="text-align:left;">
+Temporal resolution
+</th>
+<th style="text-align:left;">
+Temporal extent
 </th>
 </tr>
 </thead>
 <tbody>
-<tr>
-<td style="text-align:left;">
-<a href="https://dx.doi.org/10.5067/MODIS/MCD12Q1.061" style="     ">MCD12Q1.061</a>
-</td>
-<td style="text-align:left;">
-MODIS
-</td>
-<td style="text-align:left;">
-MODIS/Terra+Aqua Land Cover Type Yearly L3 Global 500 m SIN Grid
-</td>
-<td style="text-align:left;">
-Land cover
-</td>
-</tr>
-<tr>
-<td style="text-align:left;">
-<a href="https://dx.doi.org/10.5067/MODIS/MCD15A2H.061" style="     ">MCD15A2H.061</a>
-</td>
-<td style="text-align:left;">
-MODIS
-</td>
-<td style="text-align:left;">
-MODIS/Terra+Aqua Leaf Area Index/FPAR 8-Day L4 Global 500 m SIN Grid
-</td>
-<td style="text-align:left;">
-Leaf area index
-</td>
-</tr>
-<tr>
-<td style="text-align:left;">
-<a href="https://dx.doi.org/10.5067/MODIS/MCD15A3H.061" style="     ">MCD15A3H.061</a>
-</td>
-<td style="text-align:left;">
-MODIS
-</td>
-<td style="text-align:left;">
-MODIS/Terra+Aqua Leaf Area Index/FPAR 4-Day L4 Global 500 m SIN Grid
-</td>
-<td style="text-align:left;">
-Leaf area index
-</td>
-</tr>
 <tr>
 <td style="text-align:left;">
 <a href="https://dx.doi.org/10.5067/MODIS/MCD43A1.061" style="     ">MCD43A1.061</a>
@@ -298,11 +265,20 @@ Leaf area index
 MODIS
 </td>
 <td style="text-align:left;">
+Albedo
+</td>
+<td style="text-align:left;">
 MODIS/Terra and Aqua BRDF/Albedo Model Parameters Daily L3 Global 500 m
 SIN Grid
 </td>
 <td style="text-align:left;">
-Albedo
+500 m
+</td>
+<td style="text-align:left;">
+Daily
+</td>
+<td style="text-align:left;">
+2000-02-24 to present
 </td>
 </tr>
 <tr>
@@ -313,10 +289,19 @@ Albedo
 MODIS
 </td>
 <td style="text-align:left;">
+Albedo
+</td>
+<td style="text-align:left;">
 MODIS/Terra and Aqua BRDF/Albedo Quality Daily L3 Global 500 m SIN Grid
 </td>
 <td style="text-align:left;">
-Albedo
+500 m
+</td>
+<td style="text-align:left;">
+Daily
+</td>
+<td style="text-align:left;">
+2000-02-24 to present
 </td>
 </tr>
 <tr>
@@ -327,842 +312,19 @@ Albedo
 MODIS
 </td>
 <td style="text-align:left;">
+Albedo
+</td>
+<td style="text-align:left;">
 MODIS/Terra and Aqua Albedo Daily L3 Global 500 m SIN Grid
 </td>
 <td style="text-align:left;">
-Albedo
+500 m
 </td>
-</tr>
-<tr>
 <td style="text-align:left;">
[…]

[…]
