+++ joss-paper/paper.bib
@@ -77,8 +77,8 @@ abstract = {This paper provides a review and summary status of the research unde
          title = "The MODISTools Package: An Interface to the MODIS Land Products Subsets Web Services",
          author = "Koen Hufkens",
          year = "2022",
-         url = "https://github.com/ropensci/MODISTools",
-         doi = "10.32614/CRAN.package.MODISTools"
+         url = "https://doi.org/10.5281/zenodo.7551165",
+         doi = "10.5281/zenodo.7551165"
          }
 
 @Article{MODIStsp,
