+++ joss-paper/paper.md
@@ -84,7 +84,7 @@ Besides `modisfast`, there are several open-source R packages available for acce
 | Enables spatial subsetting at the download phase     |                        yes                        |                          yes                           |                      no                      |                        no                        |                        yes                         |                      yes                      |
 | Enables dimensional subsetting at the download phase |                        yes                        |                          yes                           |                      no                      |                       yes                        |                        yes                         |                      yes                      |
 | Maximum area size allowed for download               |                     unlimited                     |                       unlimited                        |                      NA                      |                    unlimited                     |                  200 km x 200 km                   |                   unlimited                   |
-| Website                                              | [`GitHub`](https://github.com/ptaconet/modisfast) | [`GitHub`](https://github.com/bluegreen-labs/appeears) | [`GitHub`](https://github.com/fdetsch/MODIS) | [`GitHub`](https://github.com/ropensci/MODIStsp) | [`GitHub`](https://github.com/ropensci/MODISTools) | [`GitHub`](https://github.com/r-spatial/rgee) |
+| Website                                              | [`GitHub`](https://github.com/ptaconet/modisfast) | [`GitHub`](https://github.com/bluegreen-labs/appeears) | [`GitHub`](https://github.com/fdetsch/MODIS) | [`GitHub`](https://github.com/ropensci/MODIStsp) | [`GitHub`](https://github.com/bluegreen-labs/MODISTools) | [`GitHub`](https://github.com/r-spatial/rgee) |
 | Reference                                         |                       @modisfast                            |                         @appeears                          |                      NA                      |                    @MODIStsp                     |                    @modistools                     |                     @rgee                     |
 
 Table 1: Comparison of `modisfast` with other alternatives.
