### 2024-11-12 #7343 @ptaconet
just updating the paper with a the newest version of MODISTools citation and Github repo with commit https://github.com/ptaconet/modisfast/commit/cf3169caa567bcbfe33eee3cec83b528ef24637f.
