+++ README.Rmd
@@ -60,7 +60,7 @@ This package enables to build OPeNDAP (https) URLs given input parameters such a
 
 ## Overview
 
-**`modisfast`** is an R package designed for **easy** and **fast** downloads of [**MODIS**](https://lpdaac.usgs.gov/data/get-started-data/collection-overview/missions/modis-overview/) Land products, [**VIIRS**](https://lpdaac.usgs.gov/data/get-started-data/collection-overview/missions/s-npp-nasa-viirs-overview/) Land products, and [**GPM**](https://gpm.nasa.gov/data/) (Global Precipitation Measurement Mission) Earth Observation data.
+**`modisfast`** is an R package designed for **easy** and **fast** downloads of [**MODIS**](https://lpdaac.usgs.gov/data/get-started-data/collection-overview/missions/modis-overview/) Land products, [**VIIRS**](https://lpdaac.usgs.gov/data/get-started-data/collection-overview/missions/s-npp-nasa-viirs-overview/) Land products, and [**GPM**](https://gpm.nasa.gov/data) (Global Precipitation Measurement Mission) Earth Observation data.
 
 `modisfast` uses the abilities offered by the [OPeNDAP](https://www.opendap.org/about/) framework (*Open-source Project for a Network Data Access Protocol*) to download a subset of Earth Observation data cube, along spatial, temporal or any other data dimension (depth, ...). This way, it reduces downloading time and disk usage to their minimum : no more 1° x 1° MODIS tiles with 10 bands when your region of interest is only 30 km x 30 km wide and you need 2 bands ! Moreover, `modisfast` enables parallel downloads of data.
 
@@ -261,7 +261,7 @@ There are other R packages available for accessing MODIS data. Below is a compar
 | :---------------------: | :------------------------------------------: | :-------------: |  :-------------: | :-------------: | :-------------: | :-------------------: | :-------------------: |
 | [`modisfast`](https://github.com/ptaconet/modisfast)      | MODIS, VIIRS, GPM |  :white_check_mark: |  :white_check_mark:  |   :white_check_mark:  |   :white_check_mark:  | unlimited | :white_check_mark:
 | [`appeears`](https://github.com/bluegreen-labs/appeears)  |  MODIS, VIIRS, and many others  |  :white_check_mark: |   :white_check_mark:   |   :white_check_mark: |  :white_check_mark: | unlimited | variable
-| [`MODISTools`](https://github.com/ropensci/MODISTools)  |  MODIS, VIIRS  |  :white_check_mark: |   :x:   |   :white_check_mark: |   :white_check_mark:   | 200 km x 200 km | :white_check_mark: 
+| [`MODISTools`](https://github.com/bluegreen-labs/MODISTools/)  |  MODIS, VIIRS  |  :white_check_mark: |   :x:   |   :white_check_mark: |   :white_check_mark:   | 200 km x 200 km | :white_check_mark: 
 | [`rgee`](https://github.com/r-spatial/rgee)  |  MODIS, VIIRS, GPM, and many others  |  :white_check_mark: |   :x:   |   :white_check_mark: |  :white_check_mark: | unlimited | not tested
 | [`MODIStsp`](https://github.com/ropensci/MODIStsp)   |  MODIS |  :x: |     |  :x: |   :white_check_mark:   | unlimited | NA
 | [`MODIS`](https://github.com/fdetsch/MODIS)       | MODIS  |  :x: |  :x:  |  :x:  |   :x:   | NA | NA
