### 2024-11-12 #7343 @mikemahoney218
Notes on the paper:

- Line 52&53: "long-form documentations" is an unusual phrasing. Consider "long-form tutorials" or "vignettes", or rephrase to "three pieces of long-form documentation" (though this is still a bit unusual to my ears)
- Please use braces throughout your .bib file to enforce capitalization where appropriate. Specific instances I noted:
  - Busetto, L., & Ranghetti, L. (2016): "R" should be capitalized
  - Gallagher et al 2016: Should Hyrax be capitalized in both?

Separately, and there's nothing here that you need to change, but those Hyrax citations surprised me! What an interesting way to make sure everyone gets credit.
