### 2024-11-12 6e81138
update NEWS and cran-comments
