+++ cran-comments.md
@@ -1,7 +1,5 @@
-## Resubmission
-This is a resubmission. In this version I have:
-
-* Removed the single quotes from the word MODIS in title and description in the DESCRIPTION file
+## Major release
+This is a major release, following the review of the package in the frame of the submission of a software paper in the Journal of Open Source Software (https://github.com/openjournals/joss-reviews/issues/7343)
 
 ## R CMD check results
 
