### 2024-11-12 404b902
update joss paper with comments from https://github.com/openjournals/joss-reviews/issues/7343#issuecomment-2470568943
