+++ joss-paper/paper.bib
@@ -49,7 +49,7 @@ abstract = {This paper provides a review and summary status of the research unde
 }
 
 @Manual{R,
-    title = {R: A Language and Environment for Statistical Computing},
+    title = {{R: A Language and Environment for Statistical Computing}},
     author = {{R Core Team}},
     organization = {R Foundation for Statistical Computing},
     address = {Vienna, Austria},
@@ -58,7 +58,7 @@ abstract = {This paper provides a review and summary status of the research unde
 }
 
 @article{SkofronickJackson2017,
-  title = {The Global Precipitation Measurement (GPM) Mission for Science and Society},
+  title = {{The Global Precipitation Measurement (GPM) Mission for Science and Society}},
   volume = {98},
   ISSN = {1520-0477},
   url = {10.1175/BAMS-D-15-00306.1},
@@ -83,7 +83,7 @@ abstract = {This paper provides a review and summary status of the research unde
 
 @Article{MODIStsp,
   author       = "Lorenzo Busetto and Luigi Ranghetti",
-  title        = "MODIStsp: An R Package for Preprocessing of MODIS Land Products Time Series",
+  title        = "MODIStsp: An {R} Package for Preprocessing of MODIS Land Products Time Series",
   journal      = "Computers & Geosciences",
   year         = "2016",
   volume       = "97",
@@ -107,7 +107,7 @@ abstract = {This paper provides a review and summary status of the research unde
                   Lesly Bautista and
                   Roy Yali and
                   Antony Barja},
-  title        = {{rgee: An R Package for Interacting With Google
+  title        = {{rgee: An R package for interacting with Google
                    Earth Engine}},
   month        = jul,
   year         = 2020,
@@ -120,9 +120,9 @@ abstract = {This paper provides a review and summary status of the research unde
 @software{modisfast,
   author       = {Paul Taconet and
                   Moiroux, Nicolas},
-  title        = {{modisfast: An R Package for Fast and Efficient
-                   Access to MODIS, VIIRS and GPM Earth Observation
-                   Data}},
+  title        = {{modisfast: An R package for fast and efficient
+                   access to MODIS, VIIRS and GPM Earth Observation
+                   data}},
   month        = jul,
   year         = 2024,
   publisher    = {Zenodo},
@@ -182,7 +182,7 @@ abstract = {This paper provides a review and summary status of the research unde
                   Lewis John McGibbney and
                   tsgouros and
                   ptype},
-  title        = {OPENDAP/bes: BES-3.21.0-46 for Hyrax-1.17.0},
+  title        = {{OPENDAP/bes: BES-3.21.0-46 for {H}yrax-1.17.0}},
   month        = jan,
   year         = 2024,
   publisher    = {Zenodo},
