### 2024-11-12 039009b
Increment version number to 1.0.0
