### 2024-11-12 #7343 @mikemahoney218
@ptaconet it looks like the archive is using a non-open-soirce Creative Commons license. Would you be able to update that so it matches the repository?
