### 2024-11-12 57abddc
update website and readme with new CRAN version
