### 2024-11-12 #7343 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
