### 2024-11-13 #7343 @ptaconet
Thank you so much @mikemahoney218 for the edition, and @rmendels @V-Agudetse for the review that enabled to improved the software ! I liked a lot the Github review process :)
Looking forward to submitting or reviewing in JOSS very soon !
